Add script to clear DNS cache

+++ app/scripts/restore.sh
@@ -10,8 +10,14 @@ while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
 
 sudo yes | sudo cp -rf "$HOME/Desktop/untrack/hosts" /etc/hosts
 
-sudo dscacheutil -flushcache
-sudo killall -HUP mDNSResponder
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+   sudo service dns-clean restart
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+   sudo dscacheutil -flushcache
+   sudo killall -HUP mDNSResponder
+elif [[ "$OSTYPE" == "win32" ]]; then
+   sudo ipconfig /flushdns
+fi
 
 echo 'success'
 exit
